use postgres instead of sqlite for heroku

+++ migrations/001-initial-schema.sql
@@ -1,5 +1,5 @@
 -- Up
-CREATE TABLE Instagram (id INTEGER PRIMARY KEY, url TEXT, text TEXT, post_id TEXT UNIQUE)
+CREATE TABLE Instagram (id BIGSERIAL PRIMARY KEY, url TEXT, text TEXT, post_id TEXT UNIQUE)
 
 -- Down
 DROP TABLE Instagram
